requirements.txt
anthropic>=0.39.0
rumps>=0.4.0
mcp>=1.0.0
Pillow>=10.0.0
imagehash>=4.3.0
flask>=3.0.0
pywebview>=5.0.0
rich>=13.0.0
pyobjc-framework-Cocoa>=10.0
